--- news_app_client/src/Layouts/Layouts.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import MainNavigation from './MainNavigation'
-
-const Layouts = () => {
-  return (
-    <div>
-        <MainNavigation/>
-    </div>
-  )
-}
-
-export default Layouts
