--- news_app_client/src/Layouts/MainNavigation.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import {Link} from 'react-router-dom'
-
-const MainNavigation = () => {
-  return (
-    <header>
-
-    <div>
-        
-    </div>
-    </header>
-  )
-}
-
-export default MainNavigation;
