--- news_app_client/src/Layouts/Routes.js
@@ -1,19 +1,0 @@
-// import React from 'react';
-// import { Route, Routes } from 'react-router-dom';
-// import Feeds from '../components/Feeds';
-// import MyProfile from '../components/MyProfile';
-// import Login from '../components/Login';
-// import Register from '../components/Register';
-// import CreateFeed from '../components/CreateFeed';
-// import MainNavigation from './Layouts';
-// import NavBar from './MainNavigation';
-
-// const NewsRoutes = () => {
-//   return (
-//     <div>
-       
-//     </div>
-//   )
-// }
-
-// export default NewsRoutes
