+++ news_app_client/src/components/login/login.js
@@ -1,29 +1,32 @@
 import React, { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { createUser } from '../action/user';
+import { logginUser } from '../../redux/action/user';
 import { useNavigate } from 'react-router-dom'
-import './Register.css';
 
-const Register = () => {
+const Login = () => {
+    const navigate = useNavigate();
 
-  const successMessage = useSelector(state => state.user.successMessage);
-    
-  const [userData, setUserData] = useState({
-    first_name:"",last_name:"", email: "", password: "", image: ""
-  }); 
+    const authentication = useSelector(state => state.userReducer);
 
-  const dispatch = useDispatch();
+    const [userData, setUserData] = useState({
+        email: "", password: ""
+    }); 
 
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    dispatch(createUser(userData))
-  }
+    const dispatch = useDispatch();
 
-  const navigate = useNavigate();
+    const handleSubmit = (e) => {
+        e.preventDefault();
+        dispatch(logginUser(userData))
+    }
 
-  const gotoLogin = () =>{
-    navigate('/Loggin');
-  }
+    const gotoRegister = () =>{
+        navigate('/Register');
+    }
+
+    const gotoFeeds = () =>{
+        navigate('/')
+    } 
+
   return (
     <div className='container' style={{ marginTop: '100px'}}>
         <div className="row align-items-center">
@@ -31,21 +34,39 @@ const Register = () => {
                 <form autoComplete='off' noValidate onSubmit={handleSubmit}>
                     <div className="card">
                         <div className='mx-auto col-lg-9'>
+                            {
+                                    authentication.user === null && authentication.error !== null 
+                                ? 
+                                    (
+                                        <div className="alert alert-danger" style={{ textAlign: 'center' }} role="alert">
+                                            Sorry! Username or Password not matched
+                                        </div>
+                                    ) 
+                                : 
+                                    authentication.error === null && authentication.user !== null 
+                                ? 
+                                    (
+                                        <div className="alert alert-success" style={{ textAlign: 'center' }} role="alert">
+                                            You are Successfully Logged in!
+                                        </div>
+                                    )
+                                : 
+                                    null
+                            }
+
+                            {
+                                    authentication.error === null && authentication.user !== null 
+                                ? 
+                                    (
+                                        setTimeout(() => {  gotoFeeds() }, 1500)
+                                    )
+                                : 
+                                    null
+                            }
+
                             <div><br></br></div>
-                                <h1 className='text-center'  style={{  color : "#29335c"}}>User Registration..</h1>
-                                <h4>{successMessage}</h4>
+                                <h1 className='text-center'  style={{  color : "#29335c"}}>User Loggin..</h1>
                             <div><br></br></div>
-                            <div className="form-row d-flex">
-                                <div className="form-group col-lg-6">
-                                    <label for="inputEmail4">First Name</label>
-                                    <input type="text" className="form-control" value={userData.first_name} onChange={(e)=>{ setUserData({ ...userData, first_name: e.target.value })}}/>
-                                </div>
-                                &nbsp;
-                                <div className="form-group col-lg-6">
-                                    <label for="inputPassword4">Last Name</label>
-                                    <input type="text" className="form-control" value={userData.last_name} onChange={(e)=>{ setUserData({ ...userData, last_name: e.target.value })}}/>
-                                </div>
-                            </div>
                             <div className="form-group">
                                 <label for="name">Email</label>
                                 <input type='email' className="form-control" value={userData.email} onChange={(e)=>{ setUserData({ ...userData, email: e.target.value })}}/>
@@ -53,10 +74,6 @@ const Register = () => {
                             <div className="form-group">
                                 <label for="name">Password</label>
                                 <input type='password' className="form-control" value={userData.password} onChange={(e)=>{ setUserData({ ...userData, password: e.target.value })}}/>
-                            </div>
-                            <div className="form-group">
-                                <label for="name">Image</label>
-                                <input type='file' className="form-control" value={userData.image} onChange={(e)=>{ setUserData({ ...userData, image: e.target.value })}}/>
                             </div>
                             <div><br></br></div>
                             <div className="form-group">
@@ -74,11 +91,11 @@ const Register = () => {
                             <div className="or-container"><div className="line-separator"></div> <div className="or-label">or</div><div className="line-separator"></div></div>
                             <div className="row">
                                 <div className="col-md-12 text-center">
-                                  <a className="btn btn-lg btn-google btn-block btn-outline" href='3'><img alt='' src="https://img.icons8.com/color/16/000000/google-logo.png"/>&nbsp; Signup Using Google</a>
+                                  <a className="btn btn-lg btn-google btn-block btn-outline" href='3'><img alt='' src="https://img.icons8.com/color/16/000000/google-logo.png"/>&nbsp; Loggin Using Google</a>
                                 </div>
                             </div>
                             <br/>
-                            <p className="text-inverse text-center">Already have an account ? <button className="btn btn-link" onClick={gotoLogin} data-abc="true">Login</button></p>
+                            <p className="text-inverse text-center">Don't have an account ? <button class="btn btn-link" onClick={gotoRegister} data-abc="true">Register</button></p>
                         </div>
                     </div>
                 </form>
@@ -88,4 +105,4 @@ const Register = () => {
   )
 }
 
-export default Register;
+export default Login
