+++ news_app_client/src/index.js
@@ -1,21 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
-import { BrowserRouter } from 'react-router-dom';
-import { applyMiddleware, compose, legacy_createStore as createStore } from 'redux';
-import thunk from 'redux-thunk';
-import reducers from './reducers'
-import { Provider } from 'react-redux';
+import { BrowserRouter } from 'react-router-dom'
 
-// const store = createStore(reducers, compose(applyMiddleware(thunk)))
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-    // <Provider store={store}>
       <React.StrictMode>
         <BrowserRouter>
             <App />
         </BrowserRouter>
       </React.StrictMode>
-    // </Provider>
 );
